lab bai pi 100%

+++ OOP/bai tap/tuan 2/bai 2/SoPhuc.h
@@ -1,10 +1,11 @@
 #pragma once
 #include <iostream>
+#include <math.h>
 using namespace std;
 class SoPhuc
 {
 	private:
-		int thuc, ao;
+		double thuc, ao;
 
 	public:
 		void init(int a, int b);
